fix(orders): change Order.EarnedPoints from decimal to int
EarnedPoints was declared decimal with HasPrecision(18,2), but the
entire points pipeline (CheckoutCommand computation, GreenWallet
Balance/EarnedTotal, PointTransaction.Amount) already treats points
as int, truncating before the value ever reaches this field --
the decimal precision was dead weight. Migration file only, NOT
applied to the shared dev DB (per team convention, nthung-2k5 applies
after merge).

+++ apps/backend/src/GreenEcoCommerce.Domain/Entities/Order.cs
@@ -10,7 +10,7 @@ public class Order: IHasCreatedAt
     public OrderStatusEnum Status { get; set; } = OrderStatusEnum.Pending;
     public required string DeliveryAddress { get; set; }
     public decimal DiscountAmount { get; set; }
-    public decimal EarnedPoints { get; set; }
+    public int EarnedPoints { get; set; }
     public string? CouponCode { get; set; }
     public DateTimeOffset CreatedAt { get; set; } = DateTimeOffset.UtcNow;
